requirements.txt
stockfish==3.28.0
lichess-api==0.3.1
